keep track of selected hue
This allows us to keep the selected hue when saving a black or white
color on a specific hue. So that when we close the picker and open again
we see the value/saturation screen changing. This also makes it so it is
possible to change the hue when the selected color is black or white,
previously we would need to change the color somewhere else first before
changing hue, but now we can change the hue first and then set the
value/sat pair, for example.